# api/openapi.yaml
openapi: 3.0.0
info:
  description: "Imgur's API exposes the entire Imgur infrastructure via a standardized\
    \ programmatic interface. Using Imgur's API, you can do just about anything you\
    \ can do on imgur.com, while using your programming language of choice."
  title: Imgur API
  version: 0.4.0
...
servers:
- url: https://api.imgur.com
...
paths:
...
  /3/image:
    post:
      operationId: uploadImage
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/uploadImage_request'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageResponse'
          description: Successful operation.
      security:
      - clientId: []
      - oauth2: []
      summary: Upload an image.
      tags:
      - image
      x-content-type: application/x-www-form-urlencoded
      x-accepts: application/json
  /3/image/{imageHash}:
    delete:
      operationId: deleteImage
      parameters:
      - explode: false
        in: path
        name: imageHash
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicBooleanResponse'
          description: Successful operation.
      security:
      - clientId: []
      - oauth2: []
      summary: Delete an image.
      tags:
      - image
      x-accepts: application/json
    get:
      operationId: getImage
      parameters:
      - explode: false
        in: path
        name: imageHash
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageResponse'
          description: Successful operation.
      security:
      - clientId: []
      - oauth2: []
      summary: Get information about an image.
      tags:
      - image
      x-accepts: application/json
    post:
      operationId: updateImage
      parameters:
      - explode: false
        in: path
        name: imageHash
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/updateImage_request'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicBooleanResponse'
          description: Successful operation.
      security:
      - clientId: []
      - oauth2: []
      summary: Update an image.
      tags:
      - image
      x-content-type: application/x-www-form-urlencoded
      x-accepts: application/json
  /3/image/{imageHash}/favorite:
    post:
      operationId: favoriteImage
      parameters:
      - explode: false
        in: path
        name: imageHash
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicStringResponse'
          description: Successful operation.
      security:
      - oauth2: []
      summary: Favorite an image.
      tags:
      - image
      x-accepts: application/json
components:
...
  schemas:
...
    BasicBooleanResponse:
      example:
        data: true
        success: true
        status: 0
      properties:
        data:
          type: boolean
        status:
          format: int32
          type: integer
        success:
          type: boolean
      required:
      - data
      - status
      - success
      type: object
...
    BasicStringResponse:
      example:
        data: data
        success: true
        status: 0
      properties:
        data:
          type: string
        status:
          format: int32
          type: integer
        success:
          type: boolean
      required:
      - data
      - status
      - success
      type: object
    Image:
      example:
        in_most_viral: true
        ad_type: 6
        link: link
        description: description
        section: section
        title: title
        type: type
        deletehash: deletehash
        datetime: 5
        id: id
        in_gallery: true
        vote: vote
        views: 7
        height: 5
        bandwidth: 1
        is_ad: true
        nsfw: true
        ad_url: ad_url
        tags:
        - tags
        - tags
        account_id: 0
        size: 2
        account_url: account_url
        name: name
        width: 9
        animated: true
        favorite: true
      properties:
        account_id:
          format: int32
          type: integer
        account_url:
          type: string
        ad_type:
          format: int32
          type: integer
        ad_url:
          type: string
        animated:
          type: boolean
        bandwidth:
          format: int32
          type: integer
        deletehash:
          type: string
        datetime:
          format: int32
          type: integer
        description:
          type: string
        favorite:
          type: boolean
        height:
          format: int32
          type: integer
        id:
          type: string
        in_gallery:
          type: boolean
        in_most_viral:
          type: boolean
        is_ad:
          type: boolean
        link:
          type: string
        name:
          type: string
        nsfw:
          type: boolean
        section:
          type: string
        size:
          format: int32
          type: integer
        tags:
          items:
            type: string
          type: array
        title:
          type: string
        type:
          type: string
        views:
          format: int32
          type: integer
        vote:
          type: string
        width:
          format: int32
          type: integer
      required:
      - ad_type
      - ad_url
      - animated
      - bandwidth
      - datetime
      - favorite
      - height
      - id
      - in_gallery
      - in_most_viral
      - is_ad
      - link
      - size
      - tags
      - type
      - views
      - width
      type: object
    ImageResponse:
      example:
        data:
          in_most_viral: true
          ad_type: 6
          link: link
          description: description
          section: section
          title: title
          type: type
          deletehash: deletehash
          datetime: 5
          id: id
          in_gallery: true
          vote: vote
          views: 7
          height: 5
          bandwidth: 1
          is_ad: true
          nsfw: true
          ad_url: ad_url
          tags:
          - tags
          - tags
          account_id: 0
          size: 2
          account_url: account_url
          name: name
          width: 9
          animated: true
          favorite: true
        success: true
        status: 0
      properties:
        data:
          $ref: '#/components/schemas/Image'
        status:
          format: int32
          type: integer
        success:
          type: boolean
      required:
      - data
      - status
      - success
      type: object
...
    uploadImage_request:
      properties:
        album:
          type: string
        description:
          type: string
        image:
          format: base64
          type: string
        name:
          type: string
        title:
          type: string
      required:
      - image
      type: object
    updateImage_request:
      properties:
        description:
          type: string
        title:
          type: string
      type: object
...
  securitySchemes:
    clientId:
      in: header
      name: Authorization
      type: apiKey
    oauth2:
      flows:
        implicit:
          authorizationUrl: https://api.imgur.com/oauth2/authorize
          refreshUrl: https://api.imgur.com/oauth/token
          scopes: {}
      type: oauth2